requirements.txt
qgis-plugin-ci
